Sprint 3

+++ src/main/java/ASE/rest/dto/CartPostDTO.java
@@ -9,7 +9,7 @@ public class CartPostDTO {
     private Long id;
     private List<Book> books;
     private int quantity;
-    private double prices;
+    private float prices;
     private Long userId;
 
     public Long getId() {
@@ -36,11 +36,11 @@ public class CartPostDTO {
         this.quantity=quantity;
     }
 
-    public double getPrices(){
+    public float getPrices(){
         return prices;
     }
 
-    public void setPrices(double prices){
+    public void setPrices(float prices){
         this.prices=prices;
     }
